+++ admin/src/web/controllers/socios.py
@@ -1,8 +1,9 @@
-from flask import Blueprint, session, render_template, request, redirect, flash
+from flask import Blueprint, session, render_template, request, redirect, flash, abort
 from src.core import socios
 from src.core import pagos
 from src.core import disciplinas
 from src.core import usuarios
+from src.web.helpers.permission import check_permission
 from src import exportaciones
 from src.web.controllers.validators import validator_socio
 from src.decoradores.login import login_requerido
@@ -22,30 +23,38 @@ def disciplinas_socio(id):
 @login_requerido
 def socio_index():
     """Esta funcion llama al modulo correspondiente para obtener todos los socios paginados."""
-    page = request.args.get("page", 1, type=int)
-    apellido = (
-        request.args.get("busqueda", type=str)
-        if request.args.get("busqueda", type=str) != ""
-        else None
-    )
-    tipo = (
-        request.args.get("tipo", type=str)
-        if request.args.get("tipo", type=str) != ""
-        else None
-    )
-    kwargs = {
-        "socios": socios.listar_socios(page, apellido, tipo),
-        "apellido": apellido,
-        "tipo": tipo,
-    }
-    return render_template("socios/index.html", **kwargs)
+    if check_permission(session["user"], "socio_index"):
+        page = request.args.get("page", 1, type=int)
+        apellido = (
+            request.args.get("busqueda", type=str)
+            if request.args.get("busqueda", type=str) != ""
+            else None
+        )
+        tipo = (
+            request.args.get("tipo", type=str)
+            if request.args.get("tipo", type=str) != ""
+            else None
+        )
+        kwargs = {
+            "socios": socios.listar_socios(page, apellido, tipo),
+            "apellido": apellido,
+            "tipo": tipo,
+            "usuario": usuarios.buscar_usuario_email(session["user"]),
+        }
+        return render_template("socios/index.html", **kwargs)
+    else:
+        return abort(403)
 
 
 @socio_blueprint.route("/alta-socio")
 @login_requerido
 def form_socio():
     """Esta funcion devuelve el template con un formulario para dar de alta un usuario"""
-    return render_template("socios/alta_socios.html")
+    if check_permission(session["user"], "socio_new"):
+        kwargs = {"usuario": usuarios.buscar_usuario_email(session["user"])}
+        return render_template("socios/alta_socios.html", **kwargs)
+    else:
+        return abort(403)
 
 
 @socio_blueprint.route("/<id>")
@@ -54,6 +63,7 @@ def socio_profile(id):
     """Esta funcion llama al modulo correspondiente para obtener a un socio por su id."""
     kwargs = {
         "socio": socios.buscar_socio(id),
+        "usuario": usuarios.buscar_usuario_email(session["user"]),
     }
     return render_template("socios/perfil_socio.html", **kwargs)
 
@@ -93,38 +103,44 @@ def socio_add():
 @login_requerido
 def socio_update():
     """Esta funcion llama al metodo correspondiente para modificar los datos de un socio."""
-    data_socio = {
-        "id": request.form.get("id"),
-        "nombre": request.form.get("nombre").capitalize(),
-        "apellido": request.form.get("apellido").capitalize(),
-        "email": request.form.get("email"),
-        "activo": True,
-        "tipo_documento": request.form.get("tipo_documento"),
-        "dni": request.form.get("documento"),
-        "genero": request.form.get("genero"),
-        "direccion": request.form.get("direccion"),
-        "telefono": request.form.get("telefono"),
-    }
-    validacion_datos_existentes, mensaje = socios.validar_datos_existentes(
-        data_socio["dni"], data_socio["email"], "modificacion", data_socio["id"]
-    )
-    if not validacion_datos_existentes:
-        flash(mensaje)
-        return redirect("/socios/" + data_socio["id"])
-    validacion_inputs, mensaje = validator_socio.validar_inputs(data_socio)
-    if not validacion_inputs:
-        flash(mensaje)
-        return redirect("/socios/" + data_socio["id"])
-    socios.modificar_socio(data_socio)
-    return redirect("/socios")
+    if check_permission(session["user"], "socio_update"):
+        data_socio = {
+            "id": request.form.get("id"),
+            "nombre": request.form.get("nombre").capitalize(),
+            "apellido": request.form.get("apellido").capitalize(),
+            "email": request.form.get("email"),
+            "activo": True,
+            "tipo_documento": request.form.get("tipo_documento"),
+            "dni": request.form.get("documento"),
+            "genero": request.form.get("genero"),
+            "direccion": request.form.get("direccion"),
+            "telefono": request.form.get("telefono"),
+        }
+        validacion_datos_existentes, mensaje = socios.validar_datos_existentes(
+            data_socio["dni"], data_socio["email"], "modificacion", data_socio["id"]
+        )
+        if not validacion_datos_existentes:
+            flash(mensaje)
+            return redirect("/socios/" + data_socio["id"])
+        validacion_inputs, mensaje = socios.validar_inputs(data_socio)
+        if not validacion_inputs:
+            flash(mensaje)
+            return redirect("/socios/" + data_socio["id"])
+        socio = socios.modificar_socio(data_socio)
+        return redirect("/socios")
+    else:
+        return abort(403)
 
 
 @socio_blueprint.route("/eliminar/<id>", methods=["DELETE", "GET"])
 @login_requerido
 def socio_delete(id):
     """Esta funcion llama al metodo correspondiente para eliminar un socio."""
-    socios.eliminar_socio(id)
-    return redirect("/socios")
+    if check_permission(session["user"], "socio_destroy"):
+        socio = socios.eliminar_socio(id)
+        return redirect("/socios")
+    else:
+        return abort(403)
 
 
 @socio_blueprint.route("/exportar-csv")
@@ -169,13 +185,17 @@ def exportar_pdf():
 @login_requerido
 def inscripcion_socio(id):
     """Esta funcion retorna el formulario para la inscripcion del socio a una disciplina"""
-    kwargs = {
-        "id_socio": id,
-        "disciplinas": disciplinas.todas_las_disciplinas(),
-        "categorias": disciplinas.categorias_de_cada_disciplina(),
-        "usuario": usuarios.buscar_usuario_email(session["user"]),
-    }
-    return render_template("/socios/inscripcion_socios.html", **kwargs)
+    if check_permission(session["user"], "socio_new"):
+        disciplinas2 = ["futbol", "basquet"]
+        kwargs = {
+            "id_socio": id,
+            "disciplinas": disciplinas.todas_las_disciplinas(),
+            "categorias": disciplinas.categorias_de_cada_disciplina(),
+            "usuario": usuarios.buscar_usuario_email(session["user"]),
+        }
+        return render_template("/socios/inscripcion_socios.html", **kwargs)
+    else:
+        return abort(403)
 
 
 @socio_blueprint.route("/inscripcion", methods=["POST"])
